flask/requirements.txt
Flask==1.1.1
pika==1.1.0
Werkzeug==0.15.5
pyjwt==1.7.1
